Add GUI functionality to the Portfolio Manager
This commit integrates PyQt5 to provide a GUI for the Portfolio Manager project. It includes the creation of new files necessary for setting up the GUI, such as the MainWindow, MenuBuilder, and the main application, among others. Additionally, the 'main.py' file was updated to launch the GUI, and the program's dependencies were updated to include PyQt5. Lastly, a '.gitignore' file was added for excluding 'notes' and '.idea/dataSources'.